fix(llm-usage): make kinds symmetric, guard array query serialization
Fix round 1 of Task 6 review, two follow-ups:

1. LlmUsageFilter.Kinds binds as LlmCallKind[] instead of int[], so the
   web client no longer needs a name-to-number translation layer to feed
   getLlmUsageFilters()'s enum names back into a kinds filter. Verified
   with a new HTTP-level integration test that ASP.NET Core's query-string
   enum binder accepts both the name and numeric form and that they select
   the same rows.

2. Added apps/web/src/lib/api.test.ts to guard the array-serialization fix
   from the original PR: getLlmUsage/deleteLlmUsage now have tests that
   inspect the actual wire query string (via axios.getUri) and fail if
   paramsSerializer regresses to the bracketed form ASP.NET Core's
   [FromQuery] array binder does not recognise - a regression a
   component test would not catch, and one that would silently unfilter
   the destructive bulk delete.

+++ src/Diariz.Api/Services/LlmUsageQuery.cs
@@ -66,7 +66,7 @@ public static class LlmUsageQuery
             source = source.Where(c => c.UserId != null && userIds.Contains(c.UserId.Value));
 
         if (filter.Kinds is { Length: > 0 } kinds)
-            source = source.Where(c => kinds.Contains((int)c.Kind));
+            source = source.Where(c => kinds.Contains(c.Kind));
 
         if (filter.Models is { Length: > 0 } models)
             source = source.Where(c => models.Contains(c.Model));
